vim: Explicitly set airline theme.
If the theme for airline is not set explicitly and the solarized theme
isn't available, it will silently fall back on a theme that doesn't
match.

I spent a lot of time trying to figure out why the colors for vim were
screwed up on one particular system. Turns out it was only the airline
colors that were wrong, not everything else. Installing the solarized
theme for airline fixed the issue.