general/default.pa: Disable routing for module-device-manager.
module-device-manager is auto-loaded in KDE sessions. It is used to
permit Phonon to prioritize device selection for new sound streams.

The effect of this is that my device switching scripts don't work quite
as anticipated. Applications which make use of Phonon are prioritized
that way, while other applications are not.

By explicitly loading module-device-manager with do_routing=false, we
disable the Phonon prioritization.