i3: Implement 'rofi' mode so keybindings don't hose up rofi.
Since making rofi a traditional window, I've noticed that i3 keybinds
sometimes get in the way. This change should fix that right up.

I wonder if there is a way to get rofi to show up on the active monitor
only when focus follows mouse is off in i3. I'm down a rabbit hole now!