zsh: Move functions required for env initialization into dedicated file.
Previously, zshenv sourced all files out of zsh/config/rc that had
"lib" in the name. The only purpose this served was providing pathmunge
functions.

Recently I implemented a lib function called "confirm-cmd" that zshenv
started sourcing automatically. This worked fine when logging in at
a plain old console but it broke logins via lightdm. It took me a while
to notice because I typically just lock my screen and almost never log
out.

Logins broke because ~/.xprofile is configured to source ~/.zshenv.
It choked on the fact that "confirm-cmd" was not a valid identifier
for a function.

I'm left to conclude that whatever mechanism lightdm is using to
pull in ~/.xprofile (and thereby source ~/.zshenv) must not be
actual zsh.

So, note to self: make sure all environment configuration that occurs
via ~/.zshenv is fully POSIX compliant shell.