i3: Use xss-lock for screenlocking.
The screenlocking issue described in commit 3675697 was not actually a
problem with i3lock - it was an issue with xautolock.

xautolock seems to get stuck sometimes. When it gets stuck, it will not
respond to requests to lock the screen. This was validated by running
`xautolock -locknow` from the command line, which results in:

Could not locate a running xautolock.

xss-lock appears to be a more modern solution to screenlocking than
xautolock, as it integrates with loginctl.