Implement virtual microphone for sharing local audio
During meetings or gaming sessions, audio from local applications can
play back locally in headphones, and simultaneously be casted to a
virtual microphone that is listened to by the meeting software. The
virtual microphone also has the real microphone input directed to it, so
it produces a seamless audio stream with voice and application audio.