vim/telescope: Permit find_files to follow symlinks
For note-taking, I configure multiple Syncthing directories and only
sync notes to a device that needs them. There is one main note-taking
directory with symlinks to work-related notes and personal notes.
Because I want notes to be integrated as much as they can be, Telescope
should follow symlinks when operating in the notes directory.