shell: Print a warning if "$HOME" != "$(realpath "$HOME")"
On Fedora Kinoite, $HOME contains symbolic links. This means
that on Kinoite systems:

  "$HOME" != "$(realpath "$HOME")"

This caused an impactful bug in dotbot but also presented as
a minor bash prompt bug where $HOME as the current directory
would not be abbreviated to "~".

The previous commit implemented a band-aid solution to this
problem which works for dotbot, and *probably* works for
other things. No need to take any chances, though -- we
can simply update the user's $HOME in /etc/passwd. That
is outside dotbot's scope for management but we can print
a warning when the shell launches.