vim: Fix dashboard highlighting shenanigans by disabling dashboard
I've tried a bunch of stuff and I can't *reliably* get trailing
whitespace to show up globally, except on the dashboard. The dashboard
is rendered with trailing whitespace and it looks abysmal when you start
neovim. It's a small thing but I find it quite annoying.

Even modifying the buffer content of the dashboard is not sufficient.
It appears that it is re-rendered at various times, such as when
neo-tree is opened.